requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
idna==2.9
pipenv==2018.11.26
pokemontcgsdk==2.0.0
python-decouple==3.3
pytz==2019.3
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==16.6.0
virtualenv-clone==0.5.3
